OMG FINALLYgit add .git add .git add .git add .git add . composite key works via referencing user and post, repo doesnt expose correctly will have to either find a solution or set up custom get mapping to search within the table ***WIP***

+++ src/main/java/error_search_engine/errorsearchspring/repos/PostsRepository.java
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 @Api(tags = "Address Entity")
 @RepositoryRestResource(collectionResourceRel = "posts", path = "posts")
 @CrossOrigin
-public interface PostsRepository extends CrudRepository<Posts, Integer> {
+public interface PostsRepository extends CrudRepository<Posts, Long> {
 }
